# plan-ai/backend/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiResponse_null_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"TrelloManualConnectRequest": {
				"properties": {
					"apiKey": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"apiKey",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrelloSummaryResponse": {
				"properties": {
					"totalBoards": {
						"type": "number",
						"format": "double"
					},
					"latestBoards": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"totalBoards",
					"latestBoards"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TrelloSummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TrelloSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
			"ApiResponse__success-boolean__": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"success": {
								"type": "boolean"
							}
						},
						"required": [
							"success"
						],
						"type": "object",
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
			"LinearManualConnectRequest": {
				"properties": {
					"apiKey": {
						"type": "string"
					}
				},
				"required": [
					"apiKey"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LinearSummaryResponse": {
				"properties": {
					"totalIssues": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"totalProjects": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"latestTeams": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"totalIssues",
					"totalProjects",
					"latestTeams"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_LinearSummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/LinearSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"LinearTeamItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_LinearTeamItem-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/LinearTeamItem"
						},
						"type": "array",
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"SetDefaultTeamRequest": {
				"properties": {
					"teamId": {
						"type": "string"
					}
				},
				"required": [
					"teamId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"JiraManualConnectRequest": {
				"properties": {
					"siteUrl": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"apiToken": {
						"type": "string"
					}
				},
				"required": [
					"siteUrl",
					"email",
					"apiToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"JiraSummaryResponse": {
				"properties": {
					"totalIssues": {
						"type": "string",
						"nullable": true
					},
					"totalProjects": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"latestBoards": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"totalIssues",
					"totalProjects",
					"latestBoards"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_JiraSummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/JiraSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "BUSL-1.1"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/trello/manual-connect": {
			"post": {
				"operationId": "ManualConnect",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TrelloManualConnectRequest"
							}
						}
					}
				}
			}
		},
		"/api/trello/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TrelloSummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/linear/manual-connect": {
			"post": {
				"operationId": "ManualConnect",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LinearManualConnectRequest"
							}
						}
					}
				}
			}
		},
		"/api/linear/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_LinearSummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
		"/api/linear/teams": {
			"get": {
				"operationId": "GetTeams",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_LinearTeamItem-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
		"/api/linear/default-team": {
			"post": {
				"operationId": "SetDefaultTeam",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SetDefaultTeamRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/jira/manual-connect": {
			"post": {
				"operationId": "ManualConnect",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__success-boolean__"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/JiraManualConnectRequest"
							}
						}
					}
				}
			}
		},
		"/api/jira/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_JiraSummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}